split tests for readme docs and src/lib/index.ts export into separate files
