+++ tests/multiselect.test.ts
@@ -1,5 +1,4 @@
 import { cleanup, fireEvent, render } from '@testing-library/svelte'
-import { readFileSync } from 'fs'
 import { afterEach, expect, test } from 'vitest'
 import MultiSelect from '../src/lib'
 
@@ -35,19 +34,6 @@ test(`can focus input, enter text, toggle hidden options and select an option`, 
 
   const apple_sel = getByText(`Apple`, { selector: `ul.selected > li` })
   expect(apple_sel.textContent?.trim()).toBe(`Apple`)
-})
-
-test(`readme documents all props`, () => {
-  const readme = readFileSync(`readme.md`, `utf8`)
-
-  const instance = new MultiSelect({
-    target: document.body,
-    props: { options },
-  })
-
-  for (const prop of Object.keys(instance.$$.props)) {
-    expect(readme).to.contain(prop)
-  }
 })
 
 test(`remove all button`, async () => {
